fixed bug on delete function
