+++ package.json
@@ -31,6 +31,7 @@
         "@upstatement/prettier-config": "^0.3.0"
     },
     "dependencies": {
+        "mongoose": "^6.2.0",
         "node-notifier": "^8.0.1",
         "object-path": "^0.11.5"
     }
